% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_core.R
\name{m_aggregate}
\alias{m_aggregate}
\title{m_aggregate}
\usage{
m_aggregate(merdata, types)
}
\arguments{
\item{merdata}{A \code{merdata} object}

\item{types}{A vector of types (column names) to be aggregated
e.g. "SPOL", "MLVA15"}
}
\value{
A \code{merdata} object
}
\description{
Aggregates the data in a merdata object according to columns 
specified by a vector of column labels. As an example, aggregating using 
the spoligotype column will summarise the data so each row contains a
distinct spoligotype. These aggregated columns are returned in the new
merdata object along with two augmented columns. The data is augmented with
an "ID" column where each aggregated row contains a concatenated list of
"ID" values from the original rows in the un-aggregated data. The data is
also augmented with a "COUNT" column where each aggregated row contains the
sum of the corresponding COUNT data type from the un-aggregated data.
}
\examples{
# Data aggregated to only show distinct spoligotypes
m_aggregate(monteserin, "SPOL")
# Data aggregated to show MIRU-VNTR types with distinct 
# antibiotic resistance profiles
m_aggregate(monteserin, c("MLVA15", "RES"))
}
\keyword{aggregate}
\keyword{compress}
\keyword{group}
\keyword{select}
