% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_utilities.R
\name{m_antibiotic_full_names}
\alias{m_antibiotic_full_names}
\title{m_antibiotic_full_names}
\usage{
m_antibiotic_full_names(drug_list)
}
\arguments{
\item{drug_list}{A string vector of three letter abbreviations}
}
\value{
A string vector of full antibiotic names
}
\description{
Converts antibiotic drug three letter abbreviations to full names.
Antibiotic names sourced from https://jcm.asm.org/content/abbreviations-and-conventions
}
\examples{
m_antibiotic_full_names(c("STM", "RMP"))
}
\keyword{abbreviation}
\keyword{antibiotic}
\keyword{drug}
