% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_io.R
\name{m_create}
\alias{m_create}
\title{m_create}
\usage{
m_create(details, data, validate = TRUE)
}
\arguments{
\item{details}{A data frame containing details about Format, Author, etc...}

\item{data}{A data frame containing isolate marker and resistance information}

\item{validate}{Check the resulting merdata object is valid}
}
\value{
A \code{merdata} object
}
\description{
Creates a \code{merdata} object from molecular data including header
information and data types.
}
\examples{
details <- data.frame(Format = "SMEF 1.0", Author = "John Smith", pubyear = 1900)
data <- data.frame(
  ID = c("1", "2", "3", "4", "5", "6", "7", "8"),
  COUNT = c("5", "2", "3", "7", "13", "2", "9", "4"),
  MLVA = c("25354323", "25753323", "25353322", "25353443",
           "25553323", "25353923", "25353323", "28353323"),
  NOTE = c("1", "1", "6", "4", "5", "6", "7", "7")
)
d <- m_create(details = details, data = data)
}
\keyword{merdata}
