% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_io.R
\name{m_export}
\alias{m_export}
\title{m_export}
\usage{
m_export(merdata, file, format = "SMEF")
}
\arguments{
\item{merdata}{A \code{merdata} object}

\item{file}{The path and name of the output file}

\item{format}{The output file format, SMEF or RSF}
}
\description{
Exports a \code{merdata} object to a text file in Simple 
Molecular Epidemiology Format (SMEF).
}
\examples{
# export to a file called "filename.smef" in the default SMEF format
m_export(monteserin, "filename.smef","smef")
# export to a file called "filename.rsf" in RSF format
m_export(monteserin, "filename.rsf","rsf")
}
\references{
{Tang et al. (2008). spolTools: online utilities for analyzing 
spoligotypes of the Mycobacterium tuberculosis complex. Bioinformatics,
24(20):2414-2415.}
}
\keyword{csv}
\keyword{export}
\keyword{merdata}
\keyword{rsf}
\keyword{smef}
