% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_utilities.R
\name{m_export_mst}
\alias{m_export_mst}
\title{m_export_mst}
\usage{
m_export_mst(merdata, type = "links")
}
\arguments{
\item{merdata}{A \code{merdata} object}

\item{type}{The type of data to return: "links" or "nodes"}
}
\value{
A data frame with the requested object type from the mst
}
\description{
Returns a data frame with link or node data from a merdata mst.
}
\examples{
m_export_mst(monteserin, "links")
}
\keyword{export}
\keyword{merdata}
\keyword{mst}
