% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_io.R
\name{m_get_first_type}
\alias{m_get_first_type}
\title{m_get_first_type}
\usage{
m_get_first_type(merdata, type, halt = TRUE)
}
\arguments{
\item{merdata}{A \code{merdata} object}

\item{type}{The string base type (e.g. "MLVA")}

\item{halt}{Stop on error}
}
\value{
A string containing the first instance of that type (e.g. "MLVA15")
}
\description{
Returns the first instance of a specific type. As an example if the base
type of "MLVA" is requested and there are two "MLVA" types (15, 24)
\code{m_get_first_type} will return "MLVA15".
}
\examples{
m_get_first_type(monteserin, "RES", halt = FALSE)
}
\keyword{merdata}
\keyword{type}
