% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_io.R
\name{m_get_types}
\alias{m_get_types}
\title{m_get_types}
\usage{
m_get_types(merdata, types)
}
\arguments{
\item{merdata}{A \code{merdata} object}

\item{types}{A string vector of base types (e.g. "SPOL", "MLVA")}
}
\value{
A string vector of specific types (e.g. "SPOL", "MLVA15", "MLVA24")
}
\description{
Returns a string vector containing all specific examples of the base
types requested. As an example if the base types "SPOL" and "MLVA" are
requested and the \code{merdata} object contains two "MLVA" types (15, 24)  
\code{m_get_types} will return "SPOL", "MLVA15", "MLVA24".
}
\examples{
m_get_types(monteserin, "RES")
}
\keyword{merdata}
\keyword{type}
