% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_shiny.R
\name{m_interactive}
\alias{m_interactive}
\title{m_interactive
Launch MERCAT interactive in your browser.}
\usage{
m_interactive()
}
\description{
m_interactive
Launch MERCAT interactive in your browser.
}
\examples{
\dontrun{
m_interactive()
}
}
