% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_analysis.R
\name{m_mlva}
\alias{m_mlva}
\title{m_mlva}
\usage{
m_mlva(merdata)
}
\arguments{
\item{merdata}{A \code{merdata} object}
}
\value{
A numeric matrix representing the MLVA pattern
}
\description{
m_mlva
}
\examples{
m_mlva(monteserin)
}
\keyword{MLVA}
\keyword{VNTR}
\keyword{matrix}
