% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_plot.R
\name{m_mlva_freq}
\alias{m_mlva_freq}
\title{m_mlva_freq}
\usage{
m_mlva_freq(merdata, title = "MLVA frequency", title_size = 16,
  xaxis_size = 12, yaxis_size = 12, label_size = 12, text_size = 6,
  legend_size = 12, xlab = "VNTR locus", ylab = "Repeats",
  col_high = "steelblue", col_low = "white", values = TRUE,
  legend = FALSE)
}
\arguments{
\item{merdata}{a merdata object}

\item{title}{an overall title for the plot}

\item{title_size}{main font size}

\item{xaxis_size}{X axis font size}

\item{yaxis_size}{Y axis font size}

\item{label_size}{The size of the tile labels}

\item{text_size}{The size of text for the tiles}

\item{legend_size}{the size of the legend text}

\item{xlab}{a label for the x axis}

\item{ylab}{a label for the y axis}

\item{col_high}{colour for high frequency tiles}

\item{col_low}{colour for low frequency tiles}

\item{values}{display frequency values on tiles}

\item{legend}{display a legend}
}
\description{
Plots the frequency of VNTR repeat numbers as a heat plot.
}
\examples{
m_mlva_freq(monteserin)
}
\references{
{Aminian et al. (2009). Determination of Major Lineages of Mycobacterium tuberculosis Complex using Mycobacterial Interspersed Repetitive Units. IEEE Int Conf on Bioinformatics Biomed, 1;2009:338-343.}

{Aandahl et al. (2012). A model-based Bayesian estimation of the rate of evolution of VNTR loci in Mycobacterium tuberculosis. PLoS Comput Biol. 2012;8(6):e1002573.}
}
\keyword{MIRU}
\keyword{MLVA}
\keyword{VNTR}
\keyword{merdata}
