% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_utilities.R
\name{m_mlva_names}
\alias{m_mlva_names}
\title{m_mlva_names}
\usage{
m_mlva_names(merdata)
}
\arguments{
\item{merdata}{A \code{merdata} object}
}
\value{
A string vector of MIRU-VNTR names
}
\description{
Returns MIRU-VNTR standard names according to 
\href{http://www.miru-vntrplus.org/MIRU/miruChooser.faces}{MIRU-VNTRplus}.
}
\examples{
m_mlva_names(monteserin)
}
\references{
\href{http://www.miru-vntrplus.org/MIRU/miruChooser.faces}{MIRU-VNTRplus}
}
\keyword{MIRU}
\keyword{MLVA}
\keyword{VNTR}
\keyword{merdata}
