% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_plot_graph.R
\name{m_mst}
\alias{m_mst}
\title{m_mst}
\usage{
m_mst(merdata, weights = NULL, distances = NULL, tree_seed = 1, ...)
}
\arguments{
\item{merdata}{A merdata object}

\item{weights}{A named list of weights for the different molecular types. e.g. c("MLVA" = 1, "SPOL" = 1.2); by default all markers are given weights of 1.}

\item{distances}{A named list of distance types (default c("MLVA" = "L1", "SPOL" = "spol"). Valid distances are "L1" for L1 or Manhattan distance, "hamming" for Hamming distance,}

\item{tree_seed}{A seed for the random number generation. Change the value to generate a MST with a different layout.}

\item{\dots}{Additonal plotting parameters passed to the underlying graphing function \code{m_plot_graph}}
}
\value{
A ggplot2 object
}
\description{
Plots a Minimum Spanning Tree (MST) by computing weighted distances between molecular types.
}
\examples{
m_mst(monteserin)
}
\references{
{Reyes et al. (2008). Models of deletion for visualizing bacterial variation: an application to tuberculosis spoligotypes. BMC Bioinformatics. 2008 Nov 27;9:496. doi: 10.1186/1471-2105-9-496.}
}
\keyword{MST}
\keyword{RES}
\keyword{SPOL}
\keyword{merdata}
\keyword{resistance}
\keyword{spoligotype}
