% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_plot_graph.R
\name{m_mst_nodes}
\alias{m_mst_nodes}
\title{m_mst_nodes}
\usage{
m_mst_nodes(merdata, weights = NULL, distances = NULL, tree_seed = 1)
}
\arguments{
\item{merdata}{A merdata object}

\item{weights}{A named vector of weights for the different molecular types. e.g. c("MLVA" = 1, "SPOL" = 1.2); by default all markers are given weights of 1.}

\item{distances}{A named list of distance types (default c("MLVA" = "L1", "SPOL" = "spol"). Valid distances are "L1" for L1 or Manhattan distance, "hamming" for Hamming distance,}

\item{tree_seed}{A seed for the random number generation. Change the value to generate a MST with a different layout.}
}
\value{
A data frame containing node information.
}
\description{
Prints corresponding nodes for MST plots. This function is a wrapper for \code{m_print_nodes}
}
\examples{
m_mst_nodes(monteserin)
}
\keyword{MST}
\keyword{RES}
\keyword{SPOL}
\keyword{merdata}
\keyword{resistance}
\keyword{spoligotype}
