% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_utilities.R
\name{m_oct2bin}
\alias{m_oct2bin}
\title{m_oct2bin}
\usage{
m_oct2bin(octal)
}
\arguments{
\item{octal}{A spoligotype pattern in octal format}
}
\value{
A spoligotype pattern in binary format
}
\description{
Converts an octal spoligotype pattern to binary format.
For more information see Dale, J. W., et al. "Spacer oligonucleotide typing 
of bacteria of the Mycobacterium tuberculosis complex: recommendations for 
standardised nomenclature [The Language of Our Science]." The International 
Journal of Tuberculosis and Lung Disease 5.3 (2001): 216-219.
}
\examples{
m_oct2bin("777743677760771")
}
\keyword{format}
\keyword{octal}
\keyword{spoligotype}
