% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_plot_graph.R
\name{m_plot_graph}
\alias{m_plot_graph}
\title{m_plot_graph}
\usage{
m_plot_graph(merdata, plot_type = "spolforest", weights = NULL,
  distances = NULL, tree_seed = 1, title = NULL, title_size = 16,
  node_label_size = 2, core_rad = 0.3, ring_rad = 0.05,
  spacer_rad = 0.02, ring_scale = 3, plot_scale_x = 1,
  plot_scale_y = 1, show_legend = TRUE, legend_text_size = 2,
  legend_x_location = 0.98, legend_y_location = 0.08,
  sensitive_colour = "steelblue", resistant_colour = "orange",
  unknown_colour = "grey", core_colour = "white",
  outline_colour = "black", segment_colour = "darkgrey",
  segment_label_colour = "darkgrey", show_segment_label = TRUE,
  show_node_label = TRUE, arrow_size = 0.01)
}
\arguments{
\item{merdata}{A merdata object}

\item{plot_type}{Either "spolforest" or "mst" to choose the plot type}

\item{weights}{A named vector of weights for the different molecular types. e.g. c("MLVA" = 1, "SPOL" = 1.2); by default all markers are given weights of 1.}

\item{distances}{A named list of distance types (default c("MLVA" = "L1", "SPOL" = "spol"). Valid distances are "L1" for L1 or Manhattan distance, "hamming" for Hamming distance,}

\item{tree_seed}{A seed for the random number generation. Change the value to generate a MST with a different layout.}

\item{title}{A title for the plot}

\item{title_size}{The text size for the title}

\item{node_label_size}{The text size for the node labels}

\item{core_rad}{The size of the inner nodes}

\item{ring_rad}{The base size of the outer rings}

\item{spacer_rad}{The gap between outer rings}

\item{ring_scale}{A factor to scale the rings by (uses percentage of the highest COUNT in the sample)}

\item{plot_scale_x}{A factor to scale the x-axis of the plot (1 is default)}

\item{plot_scale_y}{A factor to scale the y-axis of the plot (1 is default)}

\item{show_legend}{Display the plot resistance legend (meaningless if there are no resistance data in the merdata object)}

\item{legend_text_size}{The text size for the resistance legend}

\item{legend_x_location}{The x-axis location of the resistance legend on the plot (0-1)}

\item{legend_y_location}{The y-axis location of the resistance legend on the plot (0-1)}

\item{sensitive_colour}{The colour for drug sensitive arc segments}

\item{resistant_colour}{The colour for drug resistant arc segments}

\item{unknown_colour}{The colour for arc segments missing drug resistance information}

\item{core_colour}{The colour of the center of the nodes}

\item{outline_colour}{The colour outlining the nodes}

\item{segment_colour}{The colour of line segments between nodes}

\item{segment_label_colour}{The colour of the segment weight text (only valid for MST)}

\item{show_segment_label}{Display the segment weight (only valid for MST)}

\item{show_node_label}{Display the node ID information}

\item{arrow_size}{The size of the segment arrow (only valid for spoligoforest)}
}
\value{
A ggplot2 object
}
\description{
Plots a spoligoforest or minimum spanning tree (MST).
}
\examples{
m_plot_graph(monteserin, plot_type = "spolforest")
}
\references{
{Reyes et al. (2008). Models of deletion for visualizing bacterial variation: an application to tuberculosis spoligotypes. BMC Bioinformatics. 2008 Nov 27;9:496. doi: 10.1186/1471-2105-9-496.}
}
\keyword{MST}
\keyword{RES}
\keyword{SPOL}
\keyword{merdata}
\keyword{resistance}
\keyword{spolforest}
\keyword{spoligotype}
