% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_io.R
\name{m_read}
\alias{m_read}
\title{m_read}
\usage{
m_read(file, validate = TRUE)
}
\arguments{
\item{file}{Name (unix style) of a file in SMEF or RSF format. The function 
requires a .rsf to read RSF. The default format is SMEF.}

\item{validate}{Check to see if the file is valid.}
}
\value{
A \code{merdata} object
}
\description{
Reads a file in Simple Molecular Epidemiology Format (SMEF) 
and returns a \code{merdata} object.
}
\examples{
m_read("filename.smef")
}
\references{
{Tang et al. (2008). spolTools: online utilities for analyzing 
spoligotypes of the Mycobacterium tuberculosis complex. Bioinformatics, 
24(20):2414-2415.}
}
\keyword{data}
\keyword{input}
\keyword{smef}
