% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_analysis.R
\name{m_res}
\alias{m_res}
\title{m_res}
\usage{
m_res(merdata)
}
\arguments{
\item{merdata}{A \code{merdata} object}
}
\value{
A numeric matrix indicating the presence of antibiotic resistance
}
\description{
m_res
}
\examples{
m_res(monteserin)
}
\keyword{antibiotic}
\keyword{drug}
\keyword{matrix}
\keyword{resistance}
