% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_plot.R
\name{m_res_corr_mat}
\alias{m_res_corr_mat}
\title{m_res_corr_mat}
\usage{
m_res_corr_mat(merdata, title = "Resistance Correlation",
  title_size = 16, xaxis_size = 12, yaxis_size = 12,
  label_size = 12, text_size = 6, legend_size = 12,
  xlab = "Antibiotic", ylab = "Antibiotic", col_high = "steelblue",
  col_low = "white", col_na = "white", values = TRUE,
  legend = FALSE, symmetric = FALSE, digits = 2)
}
\arguments{
\item{merdata}{A merdata object}

\item{title}{The title of the graph}

\item{title_size}{The font size of main}

\item{xaxis_size}{The font size of the x axis}

\item{yaxis_size}{The font size of the y axis}

\item{label_size}{The size of the tile labels}

\item{text_size}{The font size for text on labels}

\item{legend_size}{The size of the legend text}

\item{xlab}{The label of the x axis}

\item{ylab}{The label of the y axis}

\item{col_high}{Colour for high frequency tiles}

\item{col_low}{Colour for low frequency tiles}

\item{col_na}{Colour for tiles without information}

\item{values}{Display correlation values on tiles}

\item{legend}{Display a legend}

\item{symmetric}{Display a symmetric graph}

\item{digits}{the significant digits to show in the plot}
}
\description{
Plots the linkage disequilibrium (LD) between pairs of antibiotic resistance frequencies. Linkage disequilibrium is computed as 
\deqn{r^2 = \frac{D_{AB}^2}{p_A(1-p_A)p_B(1-p_B)} }where \deqn{D_{AB} = p_{AB} - (p_A*p_B)} and pA and pB are the frequencies of resistance to antibiotics A and B respectively.
}
\examples{
m_res_corr_mat(monteserin)
}
\references{
{Slatkin (2008). Linkage disequilibrium-understanding the evolutionary past and mapping the medical future. Nature Reviews Genetics, 9(6):477-485, 2008.}
}
\keyword{RES}
\keyword{merdata}
\keyword{resistance}
\keyword{susceptible}
