% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_analysis.R
\name{m_res_correlation}
\alias{m_res_correlation}
\title{m_res_correlation}
\usage{
m_res_correlation(merdata)
}
\arguments{
\item{merdata}{A \code{merdata} object}
}
\value{
A numeric matrix representing the antibiotic resistance correlation
}
\description{
m_res_correlation
}
\examples{
m_res_correlation(monteserin)
}
\keyword{antibiotic}
\keyword{drug}
\keyword{matrix}
\keyword{resistance}
