% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_plot.R
\name{m_res_freq}
\alias{m_res_freq}
\title{m_res_freq}
\usage{
m_res_freq(merdata, title = "Drug resistance frequency",
  title_size = 16, xaxis_size = 12, yaxis_size = 12,
  label_size = 12, xlab = NULL, ylab = "Count", fill = "steelblue",
  orderbynames = TRUE, reverseorder = FALSE, legend = TRUE,
  legend_x = 0.5, legend_y = 1, legend_size = 4, flip = TRUE)
}
\arguments{
\item{merdata}{a merdata object}

\item{title}{title of the graph}

\item{title_size}{font size for main title}

\item{xaxis_size}{X axis font size}

\item{yaxis_size}{Y axis font size}

\item{label_size}{font size for resistance labels}

\item{xlab}{X axis label}

\item{ylab}{Y axis label}

\item{fill}{Colour of resistance bars}

\item{orderbynames}{TRUE to order by names. FALSE to order by counts.}

\item{reverseorder}{Reverse bar order}

\item{legend}{Add a legend}

\item{legend_x}{Legend X position (0-1)}

\item{legend_y}{Legend Y position (0-1)}

\item{legend_size}{Size of legend text}

\item{flip}{TRUE for horizontal bars FALSE for vertical bars}
}
\description{
Calculates a frequency histogram of drug resistant states using data of type RES.
}
\examples{
m_res_freq(monteserin)
# omit xlabel
m_res_freq(monteserin, xlab = "")
# change label size
m_res_freq(monteserin, label_size = 16)
}
\keyword{RES}
\keyword{merdata}
\keyword{resistance}
\keyword{susceptible}
