% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_analysis.R
\name{m_spol}
\alias{m_spol}
\title{m_spol}
\usage{
m_spol(merdata)
}
\arguments{
\item{merdata}{A \code{merdata} object}
}
\value{
A numeric matrix representing the spoligotype pattern
}
\description{
m_spol
}
\examples{
m_spol(monteserin)
}
\keyword{matrix}
\keyword{spoligotype}
