% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_plot_graph.R
\name{m_spolforest_nodes}
\alias{m_spolforest_nodes}
\title{m_spolforest_nodes}
\usage{
m_spolforest_nodes(merdata, tree_seed = 1)
}
\arguments{
\item{merdata}{A merdata object}

\item{tree_seed}{A seed for the random number generation.}
}
\description{
Prints corresponding nodes for a spoligoforest plot. This function is a wrapper for \code{m_print_nodes}
}
\examples{
m_spolforest_nodes(monteserin)
}
\keyword{RES}
\keyword{SPOL}
\keyword{merdata}
\keyword{resistance}
\keyword{spoligotype}
