% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_core.R
\name{m_summary}
\alias{m_summary}
\title{m_summary}
\usage{
m_summary(merdata, file = NULL, digits = 3)
}
\arguments{
\item{merdata}{A \code{merdata} object}

\item{file}{A path and name of the output file}

\item{digits}{The number of significant digits to display}
}
\description{
Writes summary statistics in "pretty" output to the screen or to a text file.
}
\examples{
# print to screen
m_summary(monteserin)
# export to a file named "output.txt"
m_summary(monteserin, "output.txt")
}
\references{
{Tang et al. (2008). spolTools: online utilities for analyzing 
spoligotypes of the Mycobacterium tuberculosis complex. 
Bioinformatics, 24(20):2414-2415.}
}
\keyword{Cluster}
\keyword{Hunter-Gaston}
\keyword{RTI}
\keyword{RTIn}
\keyword{RTIn-1}
\keyword{Size}
\keyword{genotypes}
\keyword{isolates}
\keyword{merdata}
\keyword{singleton}
\keyword{summary}
\keyword{theta}
\keyword{virtual-Heterozygosity}
