% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monteserin}
\alias{monteserin}
\title{Genotype data from 100 tuberculosis patients}
\format{A merdata object containing a list of two data frames.
\describe{
  \item{details}{details about the sample}
  \item{data}{isolate information (spoligiotype, MLVA, drug resistance)}
  ...
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/23603419}
}
\usage{
monteserin
}
\description{
Genotypes of Mycobacterium tuberculosis in patients at 
risk of drug resistance in Bolivia
Monteserin et al, 2013
}
\keyword{datasets}
