% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_core.R
\name{print.merdata}
\alias{print.merdata}
\title{Printing merdata}
\usage{
\method{print}{merdata}(x, ...)
}
\arguments{
\item{x}{A \code{merdata} object}

\item{\dots}{Further arguments passed to or from other methods.}
}
\description{
Prints each element of a merdata object using the generic function print.
}
\examples{
print(monteserin)
}
\keyword{S3}
\keyword{merdata}
\keyword{print}
