% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_core.R
\name{summary.merdata}
\alias{summary.merdata}
\title{Summary merdata}
\usage{
\method{summary}{merdata}(object, ...)
}
\arguments{
\item{object}{A \code{merdata} object}

\item{\dots}{additional arguments}
}
\value{
A list of summary statistics
}
\description{
Displays summary statistics for a \code{merdata} object.
}
\examples{
summary(monteserin)
}
\references{
{Tang et al. (2008). spolTools: online utilities for analyzing 
spoligotypes of the Mycobacterium tuberculosis complex. 
Bioinformatics, 24(20):2414-2415.}
}
\keyword{merdata}
\keyword{summary}
